trening prawie skończony
