+++ python_model/models/__init__.py
@@ -1,2 +1,2 @@
 from .model import Model
-from .cnn import CNNModel  
+from .cnn import CNNModel, HandwritingCNN
