+++ python_model/models/cnn/__init__.py
@@ -1,1 +1,2 @@
-from cnn import CNNModel
+from .cnn_model import CNNModel
+from .cnn_model_second import HandwritingCNN
