+++ python_model/models/cnn/cnn_model_second.py
@@ -1,0 +1,22 @@
+from models.model import Model
+import torch
+import torch.nn as nn
+
+class HandwritingCNN(Model):
+    def __init__(self, num_classes):
+        super().__init__()
+        self.conv1 = nn.Conv2d(1, 32, 3, padding=1)
+        self.pool1 = nn.MaxPool2d(2, 2)
+        self.conv2 = nn.Conv2d(32, 64, 3, padding=1)
+        self.pool2 = nn.MaxPool2d(2, 2)
+        self.fc1 = nn.Linear(64 * 16 * 16, 256)
+        self.fc2 = nn.Linear(256, num_classes)
+
+    def forward(self, x):
+        x = self.pool1(F.relu(self.conv1(x)))
+        x = self.pool2(F.relu(self.conv2(x)))
+        x = x.view(-1, 64 * 16 * 16)
+        x = self.relu(self.fc1(x))
+        x = self.fc2(x)
+        return x
+        return x
